+++ CHANGELOG.md
@@ -21,6 +21,8 @@ Bug fixes:
   Make `rg -vf file` where `file` is empty match everything.
 * [BUG #2177](https://github.com/BurntSushi/ripgrep/issues/2177):
   Ignore a UTF-8 BOM marker at the start of `.gitignore` (and similar files).
+* [BUG #2944](https://github.com/BurntSushi/ripgrep/pull/2944):
+  Fix a bug where the "bytes searched" in `--stats` output could be incorrect.
 
 Feature enhancements: